+++ site/MVC/Controllers/TestController.cs
@@ -47,7 +47,7 @@ namespace DFM.MVC.Controllers
 					+ getLayout(Format.WipeNotice(fakeUser, RemovalReason.NoInteraction))
 					+ getLayout(Format.WipeNotice(fakeUser, RemovalReason.NotSignedContract))
 					+ getLayout(Format.WipeNotice(fakeUser, RemovalReason.PersonAsked))
-					+ getLayout(Format.WipeCSVRecover())
+					+ getLayout(Format.WipeCSVRecover(fakeUser.Settings))
 					;
 
 			return View(result);